version that builds simons c+p successfully
