+++ regolith/builders/cpbuilder.py
@@ -2,16 +2,17 @@
 import datetime
 import time
 from copy import copy
+from nameparser import HumanName
 
 from regolith.builders.basebuilder import LatexBuilderBase
 from regolith.dates import month_to_int
 from regolith.fsclient import _id_key
 from regolith.sorters import position_key
-from regolith.chained_db import ChainDB
 from regolith.tools import (
     all_docs_from_collection,
     filter_grants,
     fuzzy_retrieval,
+    merge_collections,
     has_started,
     is_current,
 )
@@ -20,42 +21,6 @@ def is_pending(status):
     return status in "pending"
 
 
-def merge_collections(a, b, target_id):
-    """
-    merge two collections into a single merged collection
-
-    for keys that are in both collections, the value in b will be kept
-
-    Parameters
-    ----------
-    a  the inferior collection (will lose values of shared keys)
-    b  the superior collection (will keep values of shared keys)
-
-    Returns
-    -------
-    the combined collection
-    """
-    #    print(dict(b))
-    adict = {}
-    for k in a:
-        adict[k.get("_id")] = k
-    bdict = {}
-    for k in b:
-        bdict[k.get("_id")] = k
-
-    b_for_a = {}
-    for k in adict:
-        for kk, v in bdict.items():
-            if v.get(target_id, "") == k:
-                b_for_a[k] = kk
-    chained = {}
-    for k, v in b_for_a.items():
-        chained[k] = ChainDB(adict[k],
-                             bdict[v])
-#    chained.update(adict)
-    return chained
-def is_pending(sy, sm, sd):
-    return not has_started(sy, sm, sd)
 
 
 class CPBuilder(LatexBuilderBase):
@@ -91,6 +56,7 @@ class CPBuilder(LatexBuilderBase):
     def latex(self):
         """Render latex template"""
         for group in self.gtx["groups"]:
+            grp = group["_id"]
             pi = fuzzy_retrieval(
                 self.gtx["people"], ["aka", "name"], group["pi_name"]
             )
@@ -98,9 +64,8 @@ class CPBuilder(LatexBuilderBase):
             piinitialslist = [i[0] for i in pinames]
             pi['initials'] = "".join(piinitialslist).upper()
 
-            grants = list(
-                merge_collections(self.gtx["proposals"], self.gtx["grants"],
-                                  "proposal_id").values())
+            grants = merge_collections(self.gtx["proposals"], self.gtx["grants"],
+                                  "proposal_id")
             for g in grants:
                 for person in g["team"]:
                     rperson = fuzzy_retrieval(
@@ -148,12 +113,12 @@ class CPBuilder(LatexBuilderBase):
             grants = pending_grants + current_grants
             for grant in grants:
                 grant.update(
-                    award_start_date="{2}-{1}-{0}".format(
+                    award_start_date="{2}/{1}/{0}".format(
                         grant["begin_day"],
                         month_to_int(grant["begin_month"]),
                         grant["begin_year"],
                     ),
-                    award_end_date="{2}-{1}-{0}".format(
+                    award_end_date="{2}/{1}/{0}".format(
                         grant["end_day"],
                         month_to_int(grant["end_month"]),
                         grant["end_year"],
@@ -164,15 +129,16 @@ class CPBuilder(LatexBuilderBase):
             for grant in iter:
                 if grant["_id"] in badids:
                     current_grants.remove(grant)
+            piname = HumanName(pi["name"])
+            outfile = "current-pending-{}-{}".format(grp,piname.last.lower())
 
             self.render(
                 "current_pending.tex",
-                "cpp.tex",
+                outfile+".tex",
                 pi=pi,
-                #                pending=pending_grants,
                 pending=pending_grants,
                 current=current_grants,
                 pi_upper=pi["name"].upper(),
                 group=group,
             )
-            self.pdf("cpp")
+            self.pdf(outfile)
